requirements.txt
git+https://github.com/Rapptz/discord.py
pandas
python-dotenv